Move config to under persistence
